test: drop ResearchUpdate test to validate higher level component
